pyproject.toml
[project]
name = "hivetools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hive-nectar",
    "hiveengine",
    "python-dotenv",
]
